Bring the experiment up to date
For the past few days, I have been working on this experiment in my
other repository, because I am sharing code between multiple projects.

This commit brings this repo up-to-date with github.com/cbarrick/csb
(a repo which will likely be renamed in the coming months).

We're still having issues though...